requirements/base.txt
django==5.2
pillow==11.2.1
python-decouple==3.8
psycopg2-binary==2.9.10